Docker introduction
Introducing Docker for building and deploying the solution.
Configuring the solution to change debug level of a production server.
Example config for production and testing are also being committed.
Allowing for a separate location for the instance specific data to be supplied.
Updated Dockerfile and introducing a template docker-compose file
Updating readme.