Added new column to progress
Progress now has a new column to record the quality of the hours spent on a task.
The progress data point has now been changed to hours to better reflect what is being recorded to the database.
I am using a 1-5 scale for this data.
Also changed some formatting to the file and bracket choice.
Page title for Progress History changed to Time Spent