User-only screens
I have changed the design of the code to work less like the blog template to a user centric only website.
Now the login-page is the first thing seen and once logged in only the user's data can be entered.
Introduced a homepage which can be expanded upon later. Currently redirects to the goals index.
Removed the all goals sql query which is no longer needed.
Added in the login_required tag to each progress call so that login check's are performed.
Added readme info stating the parts that do not work.